[tok] Fix config.suml, reset to English
English is a better starting point for a good translation than
the current translation; if we need to look again we can check git
history.

I did some translation a number of months ago ago, but figured it would
be best to put that in the next PR to avoid getting hung on up the
translation not being perfect.